clean up api, add more specific tests

+++ sugarcrm/tests/SugarTestCurrencyUtilities.php
@@ -29,6 +29,13 @@ require('include/modules.php');
 $GLOBALS['beanList'] = $beanList;
 $GLOBALS['beanFiles'] = $beanFiles;
 
+/**
+ * SugarTestCurrencyUtilities
+ *
+ * utility class for currencies
+ *
+ * @author Monte Ohrt <mohrt@sugarcrm.com>
+ */
 class SugarTestCurrencyUtilities
 {
     private static $_createdCurrencies = array();
@@ -80,16 +87,32 @@ class SugarTestCurrencyUtilities
         return $currency;
     }
 
+    /**
+     * removeAllCreatedCurrencies
+     *
+     * remove currencies created by this test utility
+     * @return boolean true on successful removal
+     */
     public static function removeAllCreatedCurrencies()
     {
+        if(empty(self::$_createdCurrencies))
+            return true;
         $currency_ids = self::getCreatedCurrencyIds();
         $GLOBALS['db']->query(
             sprintf("DELETE FROM currencies WHERE id IN ('%s');",
             implode("','", $currency_ids))
         );
         self::$_createdCurrencies = array();
+        return true;
     }
 
+    /**
+     * getCreatedCurrencyIds
+     *
+     * get array of currency_ids created by this utility
+     *
+     * @return array list of currency_id's
+     */
     public static function getCreatedCurrencyIds()
     {
         $currency_ids = array();
